- Aerolinea -
Versión 3

+++ Aerolinea_V3/src/main/java/uv/mx/Aerolinea/EndPoint.java
@@ -17,6 +17,7 @@ import org.example.aerolinea.MostrarVueloResponse;
 import org.example.aerolinea.RealizarCompraRequest;
 import org.example.aerolinea.RealizarCompraResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.transaction.annotation.Transactional;
 import org.springframework.ws.server.endpoint.annotation.Endpoint;
 import org.springframework.ws.server.endpoint.annotation.PayloadRoot;
 import org.springframework.ws.server.endpoint.annotation.RequestPayload;
@@ -27,7 +28,7 @@ import uv.mx.Aerolinea.DTO.Vuelos;
 import uv.mx.Aerolinea.Repositorio.Icompras;
 import uv.mx.Aerolinea.Repositorio.Ivuelos;
 
-
+@Transactional
 @Endpoint
 public class EndPoint {
 	
@@ -41,26 +42,36 @@ public class EndPoint {
 	
 	@ResponsePayload
 	public AgregarVueloResponse getAgregarVuelo(@RequestPayload AgregarVueloRequest peticion ) {
-		AgregarVueloResponse agregarVuelo= new AgregarVueloResponse();
-		agregarVuelo.setMensaje("Se agrego el siguiente vuelo: \n"); 
-		agregarVuelo.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
-		agregarVuelo.setOrigen("ORIGEN: " + peticion.getOrigen());
-		agregarVuelo.setDestino("DESTINO: " + peticion.getDestino());
-		agregarVuelo.setHora("HORA: " + peticion.getHora());
-		agregarVuelo.setFecha("FECHA: " + peticion.getFecha());
-		agregarVuelo.setPrecio("PRECIO: " + peticion.getPrecio());	
+		AgregarVueloResponse agregarVuelo= new AgregarVueloResponse();						
+			
+			boolean existeVuelo;
 		
-		Vuelos vuelos = new Vuelos();
-		
-		vuelos.setIdVuelo(peticion.getIdVuelo());
-		vuelos.setOrigen(peticion.getOrigen());
-		vuelos.setDestino(peticion.getDestino());
-		vuelos.setHora(peticion.getHora());
-		vuelos.setFecha(peticion.getFecha());
-		vuelos.setPrecio(peticion.getPrecio());				
-		
-		ivuelos.save(vuelos);	
-		
+			existeVuelo=ivuelos.existsById(peticion.getIdVuelo());
+			
+			if(existeVuelo == false) {
+				agregarVuelo.setMensaje("Se agrego el siguiente vuelo: \n"); 
+				agregarVuelo.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
+				agregarVuelo.setOrigen("ORIGEN: " + peticion.getOrigen());
+				agregarVuelo.setDestino("DESTINO: " + peticion.getDestino());
+				agregarVuelo.setHora("HORA: " + peticion.getHora());
+				agregarVuelo.setFecha("FECHA: " + peticion.getFecha());
+				agregarVuelo.setPrecio("PRECIO: " + peticion.getPrecio());				
+				
+				Vuelos vuelos  = new Vuelos();		
+				
+				vuelos.setIdVuelo(peticion.getIdVuelo());
+				vuelos.setOrigen(peticion.getOrigen());
+				vuelos.setDestino(peticion.getDestino());
+				vuelos.setHora(peticion.getHora());
+				vuelos.setFecha(peticion.getFecha());
+				vuelos.setPrecio(peticion.getPrecio());
+				
+				ivuelos.save(vuelos);						
+			}else {
+				agregarVuelo.setMensaje("El Vuelo: " + peticion.getIdVuelo() + " Ya Existe !");
+			}
+			
+																							
 		return agregarVuelo;		
 	}
 		
@@ -69,31 +80,38 @@ public class EndPoint {
 		
 	@ResponsePayload
 	public MostrarVueloResponse getMostrarVuelo(@RequestPayload MostrarVueloRequest peticion ) {
-		MostrarVueloResponse mostrarVuelo= new MostrarVueloResponse();
+		MostrarVueloResponse mostrarVuelo= new MostrarVueloResponse();				
 		
-		String id = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
+		String vuelo= "";
+		vuelo = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
 		
-		String[] datosVuelo = id.split(",");
-		
-		String idVuelo = datosVuelo[0];
-		int IDVUELO = Integer.parseInt(idVuelo);
-		
-		String ORIGEN = datosVuelo[1];
-		String DESTINO = datosVuelo[2];
-		String HORA = datosVuelo[3];
-		String FECHA = datosVuelo[4];
+		if (vuelo != null) {
+			String id = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
 			
-		String precio = datosVuelo[5];
-		int PRECIO = Integer.parseInt(precio);				
-		
-		mostrarVuelo.setMensaje("Datos del vuelo seleccionado: \n");				
-		mostrarVuelo.setIdVuelo("ID_VUELO: " + IDVUELO);		
-		mostrarVuelo.setOrigen("ORIGEN: " + ORIGEN);
-		mostrarVuelo.setDestino("DESTINO: " + DESTINO);
-		mostrarVuelo.setHora("HORA: " + HORA);
-		mostrarVuelo.setFecha("FECHA: " + FECHA);
-		mostrarVuelo.setPrecio("PRECIO: " + PRECIO);
-		
+			String[] datosVuelo = id.split(",");
+			
+			String idVuelo = datosVuelo[0];
+			int IDVUELO = Integer.parseInt(idVuelo);
+			
+			String ORIGEN = datosVuelo[1];
+			String DESTINO = datosVuelo[2];
+			String HORA = datosVuelo[3];
+			String FECHA = datosVuelo[4];
+				
+			String precio = datosVuelo[5];
+			int PRECIO = Integer.parseInt(precio);				
+			
+			mostrarVuelo.setMensaje("Datos del vuelo seleccionado: \n");				
+			mostrarVuelo.setIdVuelo("ID_VUELO: " + IDVUELO);		
+			mostrarVuelo.setOrigen("ORIGEN: " + ORIGEN);
+			mostrarVuelo.setDestino("DESTINO: " + DESTINO);
+			mostrarVuelo.setHora("HORA: " + HORA);
+			mostrarVuelo.setFecha("FECHA: " + FECHA);
+			mostrarVuelo.setPrecio("PRECIO: " + PRECIO);
+		}else {
+			mostrarVuelo.setMensaje("El Vuelo: " + peticion.getIdVuelo() + " No Existe !");
+		}
+						
 		return mostrarVuelo;			
 	}
 	
@@ -101,27 +119,23 @@ public class EndPoint {
 	
 	@ResponsePayload
 	public ModificarVueloResponse getModificarVuelo(@RequestPayload ModificarVueloRequest peticion ) {
-		ModificarVueloResponse modificarVuelo= new ModificarVueloResponse();						
-		
-		modificarVuelo.setMensaje("Se modificaron los datos del siguiente vuelo: \n"); 
-		modificarVuelo.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
-		modificarVuelo.setOrigen("ORIGEN: " + peticion.getOrigen());
-		modificarVuelo.setDestino("DESTINO: " + peticion.getDestino());
-		modificarVuelo.setHora("HORA: " + peticion.getHora());
-		modificarVuelo.setFecha("FECHA: " + peticion.getFecha());
-		modificarVuelo.setPrecio("PRECIO: " + peticion.getPrecio());
-		
-		Vuelos vuelos = new Vuelos();
-		
-		vuelos.setIdVuelo(peticion.getIdVuelo());
-		vuelos.setOrigen(peticion.getOrigen());
-		vuelos.setDestino(peticion.getDestino());
-		vuelos.setHora(peticion.getHora());
-		vuelos.setFecha(peticion.getFecha());
-		vuelos.setPrecio(peticion.getPrecio());				
-		
-		ivuelos.save(vuelos);	
-		
+		ModificarVueloResponse modificarVuelo= new ModificarVueloResponse();				
+	
+		String vuelo= "";
+		vuelo = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
+		if(vuelo != null) {
+			modificarVuelo.setMensaje("Se modificaron los datos del siguiente vuelo: \n"); 
+			modificarVuelo.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
+			modificarVuelo.setOrigen("ORIGEN: " + peticion.getOrigen());
+			modificarVuelo.setDestino("DESTINO: " + peticion.getDestino());
+			modificarVuelo.setHora("HORA: " + peticion.getHora());
+			modificarVuelo.setFecha("FECHA: " + peticion.getFecha());
+			modificarVuelo.setPrecio("PRECIO: " + peticion.getPrecio());				
+			
+			ivuelos.modificarVuelo(peticion.getIdVuelo(), peticion.getOrigen(), peticion.getDestino(), peticion.getHora(), peticion.getFecha(), peticion.getPrecio());
+		}else {
+			modificarVuelo.setMensaje("El Vuelo: " + peticion.getIdVuelo() + " No Existe !");
+		}										
 		return modificarVuelo;			
 	}
 	
@@ -131,30 +145,37 @@ public class EndPoint {
 	public EliminarVueloResponse getEliminarVuelo(@RequestPayload EliminarVueloRequest peticion ) {
 		EliminarVueloResponse eliminarVuelo= new EliminarVueloResponse();
 		
-		String id = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
+		String vuelo= "";
+		vuelo = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
 		
-		String[] datosVuelo = id.split(",");
-		
-		String idVuelo = datosVuelo[0];
-		int IDVUELO = Integer.parseInt(idVuelo);
-		
-		String ORIGEN = datosVuelo[1];
-		String DESTINO = datosVuelo[2];
-		String HORA = datosVuelo[3];
-		String FECHA = datosVuelo[4];
+		if(vuelo != null) {
+			String id = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
 			
-		String precio = datosVuelo[5];
-		int PRECIO = Integer.parseInt(precio);				
-		
-		eliminarVuelo.setMensaje("Datos del vuelo eliminado: \n"); 
-		eliminarVuelo.setIdVuelo("ID_VUELO: " + IDVUELO);
-		eliminarVuelo.setOrigen("ORIGEN: " + ORIGEN);
-		eliminarVuelo.setDestino("DESTINO: " + DESTINO);
-		eliminarVuelo.setHora("HORA: " + HORA);
-		eliminarVuelo.setFecha("FECHA: " + FECHA);
-		eliminarVuelo.setPrecio("PRECIO: " + PRECIO);				
+			String[] datosVuelo = id.split(",");
+			
+			String idVuelo = datosVuelo[0];
+			int IDVUELO = Integer.parseInt(idVuelo);
+			
+			String ORIGEN = datosVuelo[1];
+			String DESTINO = datosVuelo[2];
+			String HORA = datosVuelo[3];
+			String FECHA = datosVuelo[4];
 				
-		ivuelos.deleteById(peticion.getIdVuelo());
+			String precio = datosVuelo[5];
+			int PRECIO = Integer.parseInt(precio);				
+			
+			eliminarVuelo.setMensaje("Datos del vuelo eliminado: \n"); 
+			eliminarVuelo.setIdVuelo("ID_VUELO: " + IDVUELO);
+			eliminarVuelo.setOrigen("ORIGEN: " + ORIGEN);
+			eliminarVuelo.setDestino("DESTINO: " + DESTINO);
+			eliminarVuelo.setHora("HORA: " + HORA);
+			eliminarVuelo.setFecha("FECHA: " + FECHA);
+			eliminarVuelo.setPrecio("PRECIO: " + PRECIO);				
+					
+			ivuelos.deleteById(peticion.getIdVuelo());		
+		}else {
+			eliminarVuelo.setMensaje("El Vuelo: " + peticion.getIdVuelo() + " No Existe !");
+		}		
 		
 		return eliminarVuelo;			
 	}
@@ -163,25 +184,38 @@ public class EndPoint {
 	
 	@ResponsePayload
 	public RealizarCompraResponse getRealizarCompra(@RequestPayload RealizarCompraRequest peticion ) {
-		RealizarCompraResponse realizarCompra= new RealizarCompraResponse();
-		realizarCompra.setMensaje("Datos de la compra realizada: \n");
-		realizarCompra.setIdCompra("ID_COMPRA: " + peticion.getIdCompra());
-		realizarCompra.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
-		realizarCompra.setNomCliente("NOMBRE CLIENTE: " + peticion.getNomCliente());
-		realizarCompra.setMetodoPago("METODO PAGO: " + peticion.getMetodoPago());
-		realizarCompra.setAsiento("ASIENTO: " + peticion.getAsiento());
-		
-		Compras compras = new Compras();
-		
-		compras.setIdCompra(peticion.getIdCompra());
-		compras.setIdVueloF(peticion.getIdVuelo());
-		compras.setNomCliente(peticion.getNomCliente());
-		compras.setMetodoPago(peticion.getMetodoPago());
-		compras.setAsiento(peticion.getAsiento());
+		RealizarCompraResponse realizarCompra= new RealizarCompraResponse();		
 				
-		icompras.save(compras);
+		String vuelo = "";
+		vuelo = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
 		
-		return realizarCompra;		
+		boolean existeCompra;		
+		existeCompra=icompras.existsById(peticion.getIdCompra());
+			
+			if(vuelo != null && existeCompra == false) {
+				realizarCompra.setMensaje("Datos de la compra realizada: \n");
+				realizarCompra.setIdCompra("ID_COMPRA: " + peticion.getIdCompra());
+				realizarCompra.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
+				realizarCompra.setNomCliente("NOMBRE CLIENTE: " + peticion.getNomCliente());
+				realizarCompra.setMetodoPago("METODO PAGO: " + peticion.getMetodoPago());
+				realizarCompra.setAsiento("ASIENTO: " + peticion.getAsiento());						
+			
+				Compras compras = new Compras();	
+				
+				compras.setIdCompra(peticion.getIdCompra());			
+				compras.setIdVueloF(peticion.getIdVuelo());
+				compras.setNomCliente(peticion.getNomCliente());
+				compras.setMetodoPago(peticion.getMetodoPago());
+				compras.setAsiento(peticion.getAsiento());
+							
+				icompras.save(compras);
+			}else if(vuelo == null){
+				realizarCompra.setMensaje("El Vuelo " + peticion.getIdVuelo() + " No Existe !");
+			}else if(existeCompra == true) {
+				realizarCompra.setMensaje("La Compra " + peticion.getIdCompra() + " Ya Existe !");
+			}
+		
+		return realizarCompra;	
 	}
 	
 	@PayloadRoot(localPart="MostrarCompraRequest", namespace="http://www.example.org/Aerolinea")		
@@ -190,27 +224,34 @@ public class EndPoint {
 	public MostrarCompraResponse getMostrarCompra(@RequestPayload MostrarCompraRequest peticion ) {
 		MostrarCompraResponse mostrarCompra= new MostrarCompraResponse();
 		
-		String id = icompras.mostrarDatosCompra(peticion.getIdCompra());
+		String compra = "";
+		compra = icompras.mostrarDatosCompra(peticion.getIdCompra());
 		
-		String[] datosCompra= id.split(",");
-		
-		String idCompra = datosCompra[0];
-		int IDCOMPRA = Integer.parseInt(idCompra);
-		
-		String idVuelo = datosCompra[1];
-		int IDVUELO = Integer.parseInt(idVuelo);
+		if(compra != null) {
+			String id = icompras.mostrarDatosCompra(peticion.getIdCompra());
+			
+			String[] datosCompra= id.split(",");
+			
+			String idCompra = datosCompra[0];
+			int IDCOMPRA = Integer.parseInt(idCompra);
+			
+			String idVuelo = datosCompra[1];
+			int IDVUELO = Integer.parseInt(idVuelo);
+							
+			String NOMCLIENTE = datosCompra[2];
+			String  METODOPAGO = datosCompra[3];
+			String ASIENTO = datosCompra[4];		
 						
-		String NOMCLIENTE = datosCompra[2];
-		String  METODOPAGO = datosCompra[3];
-		String ASIENTO = datosCompra[4];		
-					
-		
-		mostrarCompra.setMensaje("Datos de la compra seleccionada: \n");
-		mostrarCompra.setIdCompra("ID_COMPRA: " + IDCOMPRA);
-		mostrarCompra.setIdVuelo("ID_VUELO: " + IDVUELO);
-		mostrarCompra.setNomCliente("NOMBRE CLIENTE: " + NOMCLIENTE);
-		mostrarCompra.setMetodoPago("METODO PAGO: " + METODOPAGO);
-		mostrarCompra.setAsiento("ASIENTO: " + ASIENTO);
+			
+			mostrarCompra.setMensaje("Datos de la compra seleccionada: \n");
+			mostrarCompra.setIdCompra("ID_COMPRA: " + IDCOMPRA);
+			mostrarCompra.setIdVuelo("ID_VUELO: " + IDVUELO);
+			mostrarCompra.setNomCliente("NOMBRE CLIENTE: " + NOMCLIENTE);
+			mostrarCompra.setMetodoPago("METODO PAGO: " + METODOPAGO);
+			mostrarCompra.setAsiento("ASIENTO: " + ASIENTO);			
+		}else {
+			mostrarCompra.setMensaje("La Compra " + peticion.getIdCompra() + " No Existe !");
+		}				
 		return mostrarCompra;		
 	}
 	
@@ -219,23 +260,28 @@ public class EndPoint {
 	@ResponsePayload
 	public ModificarCompraResponse getModificarCompra(@RequestPayload ModificarCompraRequest peticion ) {
 		ModificarCompraResponse modificarCompra= new ModificarCompraResponse();
-		modificarCompra.setMensaje("Se modificaron los datos de la siguiente compra: \n");
-		modificarCompra.setIdCompra("ID_COMPRA: " + peticion.getIdCompra());
-		modificarCompra.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
-		modificarCompra.setNomCliente("NOMBRE CLIENTE: " + peticion.getNomCliente());
-		modificarCompra.setMetodoPago("METODO PAGO: " + peticion.getMetodoPago());
-		modificarCompra.setAsiento("ASIENTO: " + peticion.getAsiento());
 		
-		Compras compras = new Compras();
+		String compra = "";
+		compra = icompras.mostrarDatosCompra(peticion.getIdCompra());
 		
-		compras.setIdCompra(peticion.getIdCompra());
-		compras.setIdVueloF(peticion.getIdVuelo());		
-		compras.setNomCliente(peticion.getNomCliente());
-		compras.setMetodoPago(peticion.getMetodoPago());
-		compras.setAsiento(peticion.getAsiento());
+		String vuelo= "";
+		vuelo = ivuelos.mostrarDatosVuelo(peticion.getIdVuelo());
 		
-		icompras.save(compras);
-		
+		if(compra != null && vuelo != null) {
+			
+			modificarCompra.setMensaje("Se modificaron los datos de la siguiente compra: \n");
+			modificarCompra.setIdCompra("ID_COMPRA: " + peticion.getIdCompra());
+			modificarCompra.setIdVuelo("ID_VUELO: " + peticion.getIdVuelo());
+			modificarCompra.setNomCliente("NOMBRE CLIENTE: " + peticion.getNomCliente());
+			modificarCompra.setMetodoPago("METODO PAGO: " + peticion.getMetodoPago());
+			modificarCompra.setAsiento("ASIENTO: " + peticion.getAsiento());
+			
+			icompras.modificarCompra(peticion.getIdCompra(), peticion.getIdVuelo(), peticion.getNomCliente(), peticion.getMetodoPago(), peticion.getAsiento());			
+		}else if(compra == null){
+			modificarCompra.setMensaje("La Compra " + peticion.getIdCompra() + " No Existe !");
+		}else if(vuelo == null){
+			modificarCompra.setMensaje("El Vuelo " + peticion.getIdVuelo() + " No Existe !");
+		}
 		return modificarCompra;		
 	}
 	
@@ -245,29 +291,35 @@ public class EndPoint {
 	public CancelarCompraResponse getCancelarCompra(@RequestPayload CancelarCompraRequest peticion ) {
 		CancelarCompraResponse cancelarCompra= new CancelarCompraResponse();
 		
-		String id = icompras.mostrarDatosCompra(peticion.getIdCompra());
+		String compra = "";
+		compra = icompras.mostrarDatosCompra(peticion.getIdCompra());
 		
-		String[] datosCompra= id.split(",");
-		
-		String idCompra = datosCompra[0];
-		int IDCOMPRA = Integer.parseInt(idCompra);
-		
-		String idVuelo = datosCompra[1];
-		int IDVUELO = Integer.parseInt(idVuelo);
-						
-		String NOMCLIENTE = datosCompra[2];
-		String  METODOPAGO = datosCompra[3];
-		String ASIENTO = datosCompra[4];
-		
-		cancelarCompra.setMensaje("Datos de la compra cancelada: \n");
-		cancelarCompra.setIdCompra("ID_COMPRA: " + IDCOMPRA);
-		cancelarCompra.setIdVuelo("ID_VUELO: " + IDVUELO);
-		cancelarCompra.setNomCliente("NOMBRE CLIENTE: " + NOMCLIENTE);
-		cancelarCompra.setMetodoPago("METODO PAGO: " + METODOPAGO);
-		cancelarCompra.setAsiento("ASIENTO: " + ASIENTO);
-		
-		icompras.deleteById(peticion.getIdCompra());
-		
-		return cancelarCompra;		
+		if(compra != null) {
+			String id = icompras.mostrarDatosCompra(peticion.getIdCompra());
+			
+			String[] datosCompra= id.split(",");
+			
+			String idCompra = datosCompra[0];
+			int IDCOMPRA = Integer.parseInt(idCompra);
+			
+			String idVuelo = datosCompra[1];
+			int IDVUELO = Integer.parseInt(idVuelo);
+							
+			String NOMCLIENTE = datosCompra[2];
+			String  METODOPAGO = datosCompra[3];
+			String ASIENTO = datosCompra[4];
+			
+			cancelarCompra.setMensaje("Datos de la compra cancelada: \n");
+			cancelarCompra.setIdCompra("ID_COMPRA: " + IDCOMPRA);
+			cancelarCompra.setIdVuelo("ID_VUELO: " + IDVUELO);
+			cancelarCompra.setNomCliente("NOMBRE CLIENTE: " + NOMCLIENTE);
+			cancelarCompra.setMetodoPago("METODO PAGO: " + METODOPAGO);
+			cancelarCompra.setAsiento("ASIENTO: " + ASIENTO);
+			
+			icompras.deleteById(peticion.getIdCompra());
+		}else {
+			cancelarCompra.setMensaje("La Compra " + peticion.getIdCompra() + " No Existe !");
+		}
+		return cancelarCompra;
 	}	
 }
